Comparison code of era5 with era5t is updated.
